Fix : Separated Interface and service implementations
