+++ src/main/java/com/example/creditmarket/controller/MyPageController.java
@@ -2,7 +2,7 @@ package com.example.creditmarket.controller;
 
 import com.example.creditmarket.dto.response.FavoriteResponseDTO;
 import com.example.creditmarket.dto.response.OrderResponseDTO;
-import com.example.creditmarket.service.mypage.MyPageService;
+import com.example.creditmarket.service.MyPageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
